Delete landing-page
Originally made this file on accident, in the process of learning how to add new folders. That is, "landing-page" was intended to be a folder to be used as the publishing source for this Github Pages site. As such, this accidental file needs to be deleted before that can now be done properly.