Delete landing-page directory
Still learning Github...

Github Pages seems to want a subfolder that is also its own branch. The only way I am familiar with getting that at the moment is upon commission of a new file within a new folder. So for now, I'll just delete landing-page, and then recreate it, so that I can create a new branch for that commit. There are clearly more efficient methods to learn, but for now, this is what I know.